+++ app/models/facilities_management/journey.rb
@@ -27,7 +27,8 @@ module FacilitiesManagement
 
     FIRST_STEP_CLASS = {
       'RM3830' => FacilitiesManagement::RM3830::Journey::ChooseServices,
-      'RM6232' => FacilitiesManagement::RM6232::Journey::ChooseServices
+      'RM6232' => FacilitiesManagement::RM6232::Journey::ChooseServices,
+      'RM6378' => FacilitiesManagement::RM6378::Journey::ChooseServices
     }.freeze
   end
 end