+++ db/migrate/20250922150349_add_category_to_jurisdicitions.rb
@@ -1,0 +1,5 @@
+class AddCategoryToJurisdicitions < ActiveRecord::Migration[8.0]
+  def change
+    add_column :jurisdictions, :category, :text
+  end
+end