+++ features/support/pages/buyer_detail.rb
@@ -1,11 +1,20 @@
 module Pages
+  class SummaryRowRowSection < SitePrism::Section
+    element :key, 'dt.govuk-summary-list__key'
+    element :value, 'dd.govuk-summary-list__value'
+  end
+
   class BuyerDetail < SitePrism::Page
     section :buyer_details, '#main-content' do
-      element :Name, '#facilities_management_buyer_detail_full_name'
-      element :'Job title', '#facilities_management_buyer_detail_job_title'
-      element :'Telephone number', '#facilities_management_buyer_detail_telephone_number'
-      element :'Organisation name', '#facilities_management_buyer_detail_organisation_name'
-      element :'Organisation address', '#address-text'
+      section :'Personal details', '#buyer-details-summery--personal-details' do
+        sections :rows, SummaryRowRowSection, 'govuk-summary-list__row'
+      end
+      section :'Organisation details', '#buyer-details-summery--organisation-details' do
+        sections :rows, SummaryRowRowSection, 'govuk-summary-list__row'
+      end
+      section :'Contact preferences', '#buyer-details-summery--contact-preferences' do
+        sections :rows, SummaryRowRowSection, 'govuk-summary-list__row'
+      end
     end
 
     section :sector, '#sector-form-group' do
@@ -23,8 +32,10 @@ module Pages
       element :No, '#facilities_management_buyer_detail_contact_opt_in_false'
     end
 
+    element :address_line_1_error_message, '#organisation_address_line_1-error'
+    element :town_or_city_error_message, '#organisation_address_town-error'
     element :postcode_error_message, '#organisation_address_postcode-error'
-    element :address_drop_down, '#address-results-container'
-    element :change_address, '#change-input-2'
+    element :address_drop_down, '#organisation_address'
+    element :change_address, '#organisation-address--change-address-button'
   end
 end